# storage/api-docs/hrm-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mini ERP HRM API",
        "description": "Human Resources Management API Documentation for Mini ERP App - Includes Employee Management, Attendance, Leave, Payroll, Reimbursements, Resignations, Office Locations, and Analytics & Reporting",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/platform/hrm/employees": {
            "get": {
                "tags": [
                    "HRM Employees"
                ],
                "summary": "List all employees",
                "operationId": "5950c7af0956abe42f9de6539a7f9b83",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "department_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "designation_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "terminated",
                                "resigned"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "HRM Employees"
                ],
                "summary": "Create a new employee",
                "operationId": "a25d2c6c2d1b2092d1c175b51e1cfb9d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "Existing User ID (optional)",
                                        "type": "integer"
                                    },
                                    "first_name": {
                                        "description": "First Name",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Last Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email (required if user_id is null)",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Password (required if user_id is null)",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "department_id": {
                                        "description": "Department ID",
                                        "type": "integer"
                                    },
                                    "designation_id": {
                                        "description": "Designation ID",
                                        "type": "integer"
                                    },
                                    "emp_code": {
                                        "description": "Employee code",
                                        "type": "string"
                                    },
                                    "joining_date": {
                                        "description": "Joining date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "terminated",
                                            "resigned"
                                        ]
                                    },
                                    "nik": {
                                        "description": "National ID number",
                                        "type": "string"
                                    },
                                    "place_of_birth": {
                                        "description": "Place of birth",
                                        "type": "string"
                                    },
                                    "date_of_birth": {
                                        "description": "Date of birth",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ]
                                    },
                                    "marital_status": {
                                        "type": "string",
                                        "enum": [
                                            "single",
                                            "married",
                                            "divorced",
                                            "widowed"
                                        ]
                                    },
                                    "religion": {
                                        "description": "Religion",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "Address",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Phone number",
                                        "type": "string"
                                    },
                                    "emergency_contact_name": {
                                        "description": "Emergency contact name",
                                        "type": "string"
                                    },
                                    "emergency_contact_phone": {
                                        "description": "Emergency contact phone",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/platform/hrm/employees/{id}": {
            "get": {
                "tags": [
                    "HRM Employees"
                ],
                "summary": "Get employee details",
                "operationId": "faef458ffb4ec5ca271f74fd30ab1060",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "HRM Employees"
                ],
                "summary": "Update an employee",
                "operationId": "61858d5f905996dc2ce995d4fa21e64d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "Existing User ID",
                                        "type": "integer"
                                    },
                                    "first_name": {
                                        "description": "First Name",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Last Name",
                                        "type": "string"
                                    },
                                    "department_id": {
                                        "description": "Department ID",
                                        "type": "integer"
                                    },
                                    "designation_id": {
                                        "description": "Designation ID",
                                        "type": "integer"
                                    },
                                    "emp_code": {
                                        "description": "Employee code",
                                        "type": "string"
                                    },
                                    "joining_date": {
                                        "description": "Joining date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "terminated",
                                            "resigned"
                                        ]
                                    },
                                    "nik": {
                                        "description": "National ID number",
                                        "type": "string"
                                    },
                                    "place_of_birth": {
                                        "description": "Place of birth",
                                        "type": "string"
                                    },
                                    "date_of_birth": {
                                        "description": "Date of birth",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ]
                                    },
                                    "marital_status": {
                                        "type": "string",
                                        "enum": [
                                            "single",
                                            "married",
                                            "divorced",
                                            "widowed"
                                        ]
                                    },
                                    "religion": {
                                        "description": "Religion",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "Address",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Phone number",
                                        "type": "string"
                                    },
                                    "emergency_contact_name": {
                                        "description": "Emergency contact name",
                                        "type": "string"
                                    },
                                    "emergency_contact_phone": {
                                        "description": "Emergency contact phone",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee updated"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "HRM Employees"
                ],
                "summary": "Delete an employee",
                "operationId": "8f3f670276d73ecb60b91d007c95e7c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee deleted"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/platform/hrm/employees/{id}/enroll-face": {
            "post": {
                "tags": [
                    "HRM Employees"
                ],
                "summary": "Enroll employee face for recognition",
                "operationId": "b486f8b416b4932cb2cf17662458ef48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "face_image"
                                ],
                                "properties": {
                                    "face_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Face enrolled successfully"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/platform/hrm/employees/{id}/face-data": {
            "delete": {
                "tags": [
                    "HRM Employees"
                ],
                "summary": "Remove employee face data",
                "operationId": "fd463ca2b5b7919cd945d3a248e3172d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Face data removed"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a token from Sanctum auth",
                "name": "Token Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}